requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aniso8601==9.0.1
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
Brotli==1.0.9
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
Flask-Compress==1.10.1
Flask-RESTful==0.3.9
future==0.18.2
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.20.3
oauthlib==3.1.1
packaging==21.0
pandas==1.2.4
pkg_resources==0.0.0
plotly==4.14.3
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.4.7
Werkzeug==2.0.1
yarl==1.6.3
